+++ R/cran_results.R
@@ -1,19 +1,25 @@
-cran_results_web <- function(email, pkg, max_requests, ...) {
+cran_results_web <- function(email, pkg, include_deadline, max_requests, ...) {
   res <- NULL
   check_n_requests(email, pkg, max_requests = max_requests)
   if (!is.null(email)) {
     res_email <- read_cran_web_from_email(email, ...)
     res <- cran_checks_table(res_email)
     res <- add_other_issues(res, res_email)
-    to_fix <- fix_before_pkg_web(pkg_with_issues(res), max_requests = max_requests)
-    res <- add_fix_before(res, to_fix)
-
-}
+    to_fix <- deadline_pkg_web(
+      pkg_with_issues(res),
+      include_deadline = include_deadline,
+      max_requests = max_requests
+    )
+    res <- add_deadline(res, to_fix)
+  }
   if (!is.null(pkg)) {
     res_pkg <- read_cran_web_from_pkg(pkg, ...)
     tbl_pkg <- cran_checks_table(res_pkg)
     res_pkg <- add_other_issues(tbl_pkg, res_pkg)
-    res_pkg <- add_fix_before(res_pkg, fix_before_pkg_web(pkg, max_requests = max_requests))
+    res_pkg <- add_deadline(
+      res_pkg,
+      deadline_pkg_web(pkg, include_deadline = include_deadline, max_requests = max_requests)
+    )
     res <- rbind(res, res_pkg)
   }
   res
@@ -26,13 +32,13 @@ cran_results_crandb <- function(email, pkg, ...) {
     res_email <- read_crandb_from_email(email, ...)
     res <- cran_checks_table(res_email)
     res <- add_other_issues_crandb(res)
-    res <- add_fix_before(res, fix_before_email_crandb(email))
+    res <- add_deadline(res, deadline_email_crandb(email))
   }
   if (!is.null(pkg)) {
     res_pkg <- read_crandb_from_pkg(pkg, ...)
     tbl_pkg <- cran_checks_table(res_pkg)
     res_pkg <- add_other_issues_crandb(tbl_pkg)
-    res_pkg <- add_fix_before(res_pkg, fix_before_pkg_crandb(pkg))
+    res_pkg <- add_deadline(res_pkg, deadline_pkg_crandb(pkg))
     res <- rbind(res_pkg, res)
   }
   res
@@ -54,13 +60,17 @@ cran_results_crandb <- function(email, pkg, ...) {
 ##' @param pkg package names (character vector)
 ##' @param show columns of the data frame to show (all are shown by default)
 ##' @template src
+##' @param include_deadline if `TRUE`, the output will display the deadline set
+##'   by CRAN to fix the check results before the package gets archived. (See
+##'   Details).
 ##' @param max_requests maximum number of requests allowed to be performed,
 ##'   ignored when using `src = "crandb"`. Use `Inf` to skip this check. (See
-##'   Details)
+##'   Details).
 ##' @template dots
 ##' @template details
 ##' @return a data frame that tabulates the number of CRAN flavors that return
-##'   errors, warnings, notes, or OK for the packages.
+##'   errors, warnings, notes, or OK for the packages. See 'Details' section for
+##'   more information about the `Deadline` column.
 ##' @examples
 ##'   \dontrun{
 ##'     cran_results(pkg="MASS")
@@ -68,12 +78,16 @@ cran_results_crandb <- function(email, pkg, ...) {
 cran_results <- function(email = NULL, pkg = NULL,
                          show = c("error", "fail", "warn", "note", "ok"),
                          src = c("website", "crandb"),
+                         include_deadline = getOption("foghorn_include_deadline", TRUE),
                          max_requests = 50, ...) {
+
   show <- tolower(show)
   show <- match.arg(show, several.ok = TRUE)
-  show <- c("package", show, "has_other_issues", "fix_before")
+  show <- c("package", show, "has_other_issues", "deadline")
 
   src <- match.arg(src, c("website", "crandb"))
+
+  stopifnot(rlang::is_scalar_logical(include_deadline) && !is.na(include_deadline))
 
   if (is.null(email) && is.null(pkg)) {
     stop(
@@ -83,7 +97,13 @@ cran_results <- function(email = NULL, pkg = NULL,
   }
 
   if (identical(src, "website")) {
-    res <- cran_results_web(email, pkg, max_requests = max_requests, ...)
+    res <- cran_results_web(
+      email,
+      pkg,
+      include_deadline = include_deadline,
+      max_requests = max_requests,
+      ...
+    )
   } else if (identical(src, "crandb")) {
     res <- cran_results_crandb(email, pkg, ...)
   }