+++ NEWS.md
@@ -2,9 +2,12 @@
 
 ## New Features
 
-* The output of `cran_results()` and `summary_cran_results()` now include the
-  deadline fixed by CRAN. This deadline is the date by which requested changes
-  need to be implemented before the package gets archived.
+* The output of `cran_results()` and `summary_cran_results()` can now include
+  the deadline fixed by CRAN. This deadline is the date by which requested
+  changes need to be implemented before the package gets archived. By default,
+  this is turned off. To see any deadline that might have been set by CRAN
+  Maintainers, set the `include_deadline` argument to `TRUE`. The value for this
+  argument can also be set using the local option `foghorn_include_deadline`.
 * The `cran_results()` function gains the `max_requests` argument that limits
   the number of requests that are performed against the CRAN website in a single
   query.